Remove MAX_TREE_DEPTH limit, use varints for depth/TTL
Breaking protocol change that removes artificial limits on tree depth,
enabling continental-scale mesh networks (e.g., Oslo to Rome via LoRa).

Changes:
- Remove MAX_TREE_DEPTH constant (was 127)
- Change depth, max_depth, ttl from u8 to u32 with varint wire encoding
- Dynamic TTL: max(255, max_depth * 3) instead of fixed DEFAULT_TTL
- Bounce-back restores original TTL (not reset), since it's retry not restart
- Update recently_forwarded to store TTL for restoration
- Update wire format: Pulse depth/max_depth and Routed ttl now varint-encoded

Rationale: Physical limits (latency, TTL exhaustion) naturally bound tree
depth. With MAX_CHILDREN=12, depth ≈ log₁₂(N). The protocol should not
impose additional artificial limits.

Wire efficiency: Shallow trees (depth < 128) still use 1 byte per field.

Reviewed by embedded-rust-engineer and protocol-security-architect agents.
Note: Security review identified potential max_depth inflation attack
(attacker could claim high max_depth to inflate network TTL values).
This is accepted as a known tradeoff - physics bounds the actual impact.

+++ darktree/src/debug.rs
@@ -144,7 +144,7 @@ pub enum DebugEvent {
         payload_hash: [u8; 4],
         msg_type: u8,
         dest_addr: u32,
-        ttl: u8,
+        ttl: u32,
     },
     /// Routed message forwarded to next hop.
     /// `direction` is "up" (to parent) or "down" (to child/shortcut).
@@ -153,7 +153,7 @@ pub enum DebugEvent {
         msg_type: u8,
         dest_addr: u32,
         next_hop: NodeId,
-        ttl: u8,
+        ttl: u32,
         my_keyspace: (u32, u32),
         direction: &'static str,
     },
@@ -217,7 +217,7 @@ pub enum DebugEvent {
         payload_hash: [u8; 4],
         msg_type: u8,
         dest_addr: u32,
-        ttl: u8,
+        ttl: u32,
         has_route: bool,
     },
     /// Outgoing queue rejected a routed message (queue full).
